mypt-setting-api/requirements.txt
django
djangorestframework
markdown
django-filter
pyjwt
mysqlclient
pycryptodome
pycryptodomex
numpy
pandas
redis
python-decouple
requests
selenium
webdriver-manager
cron-schedule
